Move findDeclaration into a more logical place and minor restructuring
