+++ server/src/main/kotlin/org/javacs/kt/CompiledFile.kt
@@ -4,6 +4,7 @@ import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 import org.javacs.kt.compiler.CompilationKind
 import org.javacs.kt.position.changedRegion
+import org.javacs.kt.position.location
 import org.javacs.kt.position.position
 import org.javacs.kt.util.findParent
 import org.javacs.kt.util.nullResult
@@ -11,11 +12,13 @@ import org.javacs.kt.util.toPath
 import org.jetbrains.kotlin.descriptors.DeclarationDescriptor
 import org.jetbrains.kotlin.descriptors.ModuleDescriptor
 import org.jetbrains.kotlin.lexer.KtTokens
+import org.jetbrains.kotlin.js.resolve.diagnostics.findPsi
 import org.jetbrains.kotlin.psi.*
 import org.jetbrains.kotlin.psi.psiUtil.parentsWithSelf
 import org.jetbrains.kotlin.resolve.BindingContext
 import org.jetbrains.kotlin.resolve.scopes.LexicalScope
 import org.jetbrains.kotlin.types.KotlinType
+import org.eclipse.lsp4j.Location
 import java.nio.file.Paths
 
 class CompiledFile(
@@ -169,6 +172,25 @@ class CompiledFile(
         val oldCursor = oldOffset(cursor)
         val psi = parse.findElementAt(oldCursor) ?: return nullResult("Couldn't find anything at ${describePosition(cursor)}")
         return psi.findParent<KtElement>()
+    }
+
+
+    /**
+     * Find the declaration of the element at the cursor. Only works if the element at the cursor is a reference.
+     */
+    fun findDeclaration(cursor: Int): Pair<KtNamedDeclaration, Location>? {
+        val (_, target) = referenceAtPoint(cursor) ?: return null
+        val psi = target.findPsi()
+
+        return if (psi is KtNamedDeclaration) {
+            psi.nameIdentifier?.let {
+                location(it)?.let { location ->
+                    Pair(psi, location)
+                }
+            }
+        } else {
+            null
+        }
     }
 
     /**
